requirements.txt
streamlit
moviepy
opencv-python
SpeechRecognition
Pillow